refactor(utils): move natural_cmp and quality-tier helpers into shared
Issue: the natural (numeric-aware) comparator lived privately in the
backend sort processor, and the quality-token ranking lived privately
in the dedup module. The upcoming quality filter field needs the
ranking in shared, and UI lists cannot reuse the natural ordering at
all while it is backend-private.

Change: natural_cmp (with its leading-zero tie-break) plus
token_quality/quality_tokens/quality_rank move to
shared/src/utils/string_utils.rs as public helpers; the sort
processor imports natural_cmp from shared and drops its local copy.
Comparator and quality-rank unit tests move along into the shared
test module. No behavior change.

Benefit: one canonical implementation usable by backend processing,
the shared filter DSL, and the WASM frontend alike.
